Updater: sync 2 sources of Red Hat's security data
Due to 2 different sources of Red Hat security data we have to make sure
we don't miss any advisory which could be cause by race condition of
those 2 sources.

This commit updates Red Hat's updater and it should not miss any RHSA.
The advisories which can be missing in rhsa2cpe mapping file are stored
in keyvalue DB table and they are processed in next updater run.